+++ Sources/Sentry/SentryHub.m
@@ -26,25 +26,24 @@
 
 NS_ASSUME_NONNULL_BEGIN
 
-@interface SentryHub () <SentryLoggerDelegate>
+@interface SentryHubInternal () <SentryLoggerDelegate>
 
-@property (nullable, atomic, strong) SentryClient *client;
+@property (nullable, atomic, strong) SentryClientInternal *client;
 @property (nullable, nonatomic, strong) SentryScope *scope;
 @property (nonatomic) SentryDispatchQueueWrapper *dispatchQueue;
 @property (nonatomic, strong) SentryCrashWrapper *crashWrapper;
 @property (nonatomic, strong) NSMutableSet<NSString *> *installedIntegrationNames;
 @property (nonatomic) NSUInteger errorsBeforeSession;
 @property (nonatomic, weak) id<SentrySessionListener> sessionListener;
-@property (nonatomic, strong) SentryLogger *logger;
 
 @end
 
-@implementation SentryHub {
+@implementation SentryHubInternal {
     NSObject *_sessionLock;
     NSObject *_integrationsLock;
 }
 
-- (instancetype)initWithClient:(nullable SentryClient *)client
+- (instancetype)initWithClient:(nullable SentryClientInternal *)client
                       andScope:(nullable SentryScope *)scope
 {
     return [self initWithClient:client
@@ -54,7 +53,7 @@ NS_ASSUME_NONNULL_BEGIN
 }
 
 /** Internal constructor for testing */
-- (instancetype)initWithClient:(nullable SentryClient *)client
+- (instancetype)initWithClient:(nullable SentryClientInternal *)client
                       andScope:(nullable SentryScope *)scope
                andCrashWrapper:(SentryCrashWrapper *)crashWrapper
               andDispatchQueue:(SentryDispatchQueueWrapper *)dispatchQueue
@@ -175,7 +174,7 @@ NS_ASSUME_NONNULL_BEGIN
     SENTRY_LOG_DEBUG(@"A cached session was found.");
 
     // Make sure there's a client bound.
-    SentryClient *client = self.client;
+    SentryClientInternal *client = self.client;
     if (client == nil) {
         SENTRY_LOG_DEBUG(@"No client bound.");
         return;
@@ -205,7 +204,7 @@ NS_ASSUME_NONNULL_BEGIN
         return;
     }
 
-    SentryClient *client = self.client;
+    SentryClientInternal *client = self.client;
     if (client.options.diagnosticLevel == kSentryLevelDebug) {
         SENTRY_LOG_DEBUG(@"Capturing session with status: %@",
             [self createSessionDebugString:SENTRY_UNWRAP_NULLABLE(SentrySession, session)]);
@@ -242,7 +241,7 @@ NS_ASSUME_NONNULL_BEGIN
 {
     event.isFatalEvent = YES;
 
-    SentryClient *client = self.client;
+    SentryClientInternal *client = self.client;
     if (client == nil) {
         return;
     }
@@ -272,7 +271,7 @@ NS_ASSUME_NONNULL_BEGIN
     // We treat fatal app hang events similar to crashes.
     event.isFatalEvent = YES;
 
-    SentryClient *_Nullable client = _client;
+    SentryClientInternal *_Nullable client = _client;
     if (client == nil) {
         return;
     }
@@ -319,7 +318,7 @@ NS_ASSUME_NONNULL_BEGIN
     // When a user calls finish on a transaction, which calls captureTransaction, the calling thread
     // here could be the main thread, which we only want to block as long as required. Therefore, we
     // capture the transaction on a background thread.
-    __weak SentryHub *weakSelf = self;
+    __weak SentryHubInternal *weakSelf = self;
     [self.dispatchQueue dispatchAsyncWithBlock:^{
         [weakSelf captureEvent:transaction
                           withScope:scope
@@ -336,7 +335,7 @@ NS_ASSUME_NONNULL_BEGIN
         return;
     }
 
-    SentryClient *client = self.client;
+    SentryClientInternal *client = self.client;
     if (client != nil) {
         [client saveCrashTransaction:transaction withScope:self.scope];
     }
@@ -356,7 +355,7 @@ NS_ASSUME_NONNULL_BEGIN
                   withScope:(SentryScope *)scope
     additionalEnvelopeItems:(NSArray<SentryEnvelopeItem *> *)additionalEnvelopeItems
 {
-    SentryClient *client = self.client;
+    SentryClientInternal *client = self.client;
     if (client != nil) {
         return [client captureEvent:event
                           withScope:scope
@@ -480,7 +479,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 - (SentryId *)captureMessage:(NSString *)message withScope:(SentryScope *)scope
 {
-    SentryClient *client = self.client;
+    SentryClientInternal *client = self.client;
     if (client != nil) {
         return [client captureMessage:message withScope:scope];
     }
@@ -495,7 +494,7 @@ NS_ASSUME_NONNULL_BEGIN
 - (SentryId *)captureError:(NSError *)error withScope:(SentryScope *)scope
 {
     SentrySession *currentSession = _session;
-    SentryClient *client = self.client;
+    SentryClientInternal *client = self.client;
     if (client != nil) {
         if (currentSession != nil) {
             return [client captureError:error
@@ -517,7 +516,7 @@ NS_ASSUME_NONNULL_BEGIN
 - (SentryId *)captureException:(NSException *)exception withScope:(SentryScope *)scope
 {
     SentrySession *currentSession = _session;
-    SentryClient *client = self.client;
+    SentryClientInternal *client = self.client;
     if (client != nil) {
         if (currentSession != nil) {
             return [client captureException:exception
@@ -533,7 +532,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 - (void)captureFeedback:(SentryFeedback *)feedback
 {
-    SentryClient *client = self.client;
+    SentryClientInternal *client = self.client;
     if (client != nil) {
         [client captureFeedback:feedback withScope:self.scope];
     }
@@ -543,7 +542,7 @@ NS_ASSUME_NONNULL_BEGIN
                       withEventId:(NSString *)feedbackEventId
                       attachments:(NSArray<SentryAttachment *> *)feedbackAttachments
 {
-    SentryClient *client = self.client;
+    SentryClientInternal *client = self.client;
     if (client != nil) {
         [client captureSerializedFeedback:serializedFeedback
                               withEventId:feedbackEventId
@@ -570,12 +569,12 @@ NS_ASSUME_NONNULL_BEGIN
     [self.scope addBreadcrumb:SENTRY_UNWRAP_NULLABLE(SentryBreadcrumb, nullableCrumb)];
 }
 
-- (nullable SentryClient *)getClient
+- (nullable SentryClientInternal *)getClient
 {
     return self.client;
 }
 
-- (void)bindClient:(nullable SentryClient *)client
+- (void)bindClient:(nullable SentryClientInternal *)client
 {
     self.client = client;
 }
@@ -584,7 +583,7 @@ NS_ASSUME_NONNULL_BEGIN
 {
     @synchronized(self) {
         if (_scope == nil) {
-            SentryClient *client = self.client;
+            SentryClientInternal *client = self.client;
             if (client != nil) {
                 _scope = [[SentryScope alloc] initWithMaxBreadcrumbs:client.options.maxBreadcrumbs];
             } else {
@@ -600,7 +599,7 @@ NS_ASSUME_NONNULL_BEGIN
 - (void)configureScope:(void (^)(SentryScope *scope))callback
 {
     SentryScope *scope = self.scope;
-    SentryClient *client = _client;
+    SentryClientInternal *client = _client;
     if (client != nil && scope != nil) {
         callback(scope);
     }
@@ -692,7 +691,7 @@ NS_ASSUME_NONNULL_BEGIN
  */
 - (void)storeEnvelope:(SentryEnvelope *)envelope
 {
-    SentryClient *client = self.client;
+    SentryClientInternal *client = self.client;
     if (client == nil) {
         return;
     }
@@ -704,7 +703,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 - (void)captureEnvelope:(SentryEnvelope *)envelope
 {
-    SentryClient *client = self.client;
+    SentryClientInternal *client = self.client;
     if (client == nil) {
         return;
     }
@@ -814,7 +813,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 - (void)flush:(NSTimeInterval)timeout
 {
-    SentryClient *client = self.client;
+    SentryClientInternal *client = self.client;
     if (client != nil) {
         [client flush:timeout];
     }
@@ -842,7 +841,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 - (void)captureLog:(SentryLog *)log
 {
-    SentryClient *client = self.client;
+    SentryClientInternal *client = self.client;
     if (client != nil) {
 #if SENTRY_TARGET_REPLAY_SUPPORTED
         NSMutableDictionary<NSString *, SentryStructuredLogAttribute *> *mutableAttributes =