+++ Sources/Sentry/include/SentryClient.h
@@ -8,15 +8,15 @@
 
 @class SentryEvent;
 @class SentryFeedback;
-@class SentryId;
 @class SentryOptions;
 @class SentryScope;
+@class SentryId;
 @class SentryTransaction;
 @class SentryLog;
 
 NS_ASSUME_NONNULL_BEGIN
 
-@interface SentryClient : NSObject
+@interface SentryClientInternal : NSObject
 SENTRY_NO_INIT
 
 @property (nonatomic, assign, readonly) BOOL isEnabled;