+++ Sources/Sentry/include/SentryHub.h
@@ -8,7 +8,7 @@
 #import SENTRY_HEADER(SentrySpanProtocol)
 
 @class SentryBreadcrumb;
-@class SentryClient;
+@class SentryClientInternal;
 @class SentryEvent;
 @class SentryFeedback;
 @class SentryId;
@@ -19,10 +19,10 @@
 @class SentryLogger;
 
 NS_ASSUME_NONNULL_BEGIN
-@interface SentryHub : NSObject
+@interface SentryHubInternal : NSObject
 SENTRY_NO_INIT
 
-- (instancetype)initWithClient:(SentryClient *_Nullable)client
+- (instancetype)initWithClient:(SentryClientInternal *_Nullable)client
                       andScope:(SentryScope *_Nullable)scope;
 
 /**
@@ -193,7 +193,7 @@ SENTRY_NO_INIT
 /**
  * Returns a client if there is a bound client on the Hub.
  */
-- (SentryClient *_Nullable)getClient;
+- (SentryClientInternal *_Nullable)getClient;
 
 /**
  * Returns either the current scope or a new one if it was @c nil .
@@ -208,7 +208,7 @@ SENTRY_NO_INIT
 /**
  * Binds a different client to the hub.
  */
-- (void)bindClient:(SentryClient *_Nullable)client;
+- (void)bindClient:(SentryClientInternal *_Nullable)client;
 
 /**
  * Checks if integration is activated.