+++ Sources/Swift/Helper/SentrySDK.swift
@@ -30,7 +30,10 @@ import Foundation
         if !SentrySDKInternal.isEnabled {
             SentrySDKLog.fatal("Logs called before SentrySDK.start() will be dropped.")
         }
-        return SentrySDKInternal.currentHub().logger
+        // Temp until we figure out how to access the logger from the helper
+        // swiftlint:disable force_cast
+        return SentrySDKInternal.currentHub().value(forKey: "logger") as! SentryLogger
+        // swiftlint:enable force_cast
     }
     
     /// Inits and configures Sentry (`SentryHub`, `SentryClient`) and sets up all integrations. Make sure to
@@ -56,7 +59,7 @@ import Foundation
     /// - returns: The `SentryId` of the event or `SentryId.empty` if the event is not sent.
     @objc(captureEvent:)
     @discardableResult public static func capture(event: Event) -> SentryId {
-        return SentrySDKInternal.capture(event: event).sentryId
+        return SentrySDKInternal.capture(event: event)
     }
     
     /// Captures a manually created event and sends it to Sentry. Only the data in this scope object will
@@ -66,7 +69,7 @@ import Foundation
     /// - returns: The `SentryId` of the event or `SentryId.empty` if the event is not sent.
     @objc(captureEvent:withScope:)
     @discardableResult public static func capture(event: Event, scope: Scope) -> SentryId {
-        return SentrySDKInternal.capture(event: event, scope: scope).sentryId
+        return SentrySDKInternal.capture(event: event, scope: scope)
     }
     
     /// Captures a manually created event and sends it to Sentry. Maintains the global scope but mutates
@@ -76,7 +79,7 @@ import Foundation
     /// - returns: The `SentryId` of the event or `SentryId.empty` if the event is not sent.
     @objc(captureEvent:withScopeBlock:)
     @discardableResult public static func capture(event: Event, block: @escaping (Scope) -> Void) -> SentryId {
-        return SentrySDKInternal.capture(event: event, block: block).sentryId
+        return SentrySDKInternal.capture(event: event, block: block)
     }
     
     // MARK: - Transaction Management
@@ -141,7 +144,7 @@ import Foundation
     /// - returns: The `SentryId` of the event or `SentryId.empty` if the event is not sent.
     @objc(captureError:)
     @discardableResult public static func capture(error: Error) -> SentryId {
-        return SentrySDKInternal.capture(error: error).sentryId
+        return SentrySDKInternal.capture(error: error)
     }
     
     /// Captures an error event and sends it to Sentry. Only the data in this scope object will be added
@@ -151,7 +154,7 @@ import Foundation
     /// - returns: The `SentryId` of the event or `SentryId.empty` if the event is not sent.
     @objc(captureError:withScope:)
     @discardableResult public static func capture(error: Error, scope: Scope) -> SentryId {
-        return SentrySDKInternal.capture(error: error, scope: scope).sentryId
+        return SentrySDKInternal.capture(error: error, scope: scope)
     }
     
     /// Captures an error event and sends it to Sentry. Maintains the global scope but mutates scope data
@@ -161,7 +164,7 @@ import Foundation
     /// - returns: The `SentryId` of the event or `SentryId.empty` if the event is not sent.
     @objc(captureError:withScopeBlock:)
     @discardableResult public static func capture(error: Error, block: @escaping (Scope) -> Void) -> SentryId {
-        return SentrySDKInternal.capture(error: error, block: block).sentryId
+        return SentrySDKInternal.capture(error: error, block: block)
     }
     
     // MARK: - Exception Capture
@@ -171,7 +174,7 @@ import Foundation
     /// - returns: The `SentryId` of the event or `SentryId.empty` if the event is not sent.
     @objc(captureException:)
     @discardableResult public static func capture(exception: NSException) -> SentryId {
-        return SentrySDKInternal.capture(exception: exception).sentryId
+        return SentrySDKInternal.capture(exception: exception)
     }
     
     /// Captures an exception event and sends it to Sentry. Only the data in this scope object will be
@@ -181,7 +184,7 @@ import Foundation
     /// - returns: The `SentryId` of the event or `SentryId.empty` if the event is not sent.
     @objc(captureException:withScope:)
     @discardableResult public static func capture(exception: NSException, scope: Scope) -> SentryId {
-        return SentrySDKInternal.capture(exception: exception, scope: scope).sentryId
+        return SentrySDKInternal.capture(exception: exception, scope: scope)
     }
     
     /// Captures an exception event and sends it to Sentry. Maintains the global scope but mutates scope
@@ -191,7 +194,7 @@ import Foundation
     /// - returns: The `SentryId` of the event or `SentryId.empty` if the event is not sent.
     @objc(captureException:withScopeBlock:)
     @discardableResult public static func capture(exception: NSException, block: @escaping (Scope) -> Void) -> SentryId {
-        return SentrySDKInternal.capture(exception: exception, block: block).sentryId
+        return SentrySDKInternal.capture(exception: exception, block: block)
     }
     
     // MARK: - Message Capture
@@ -201,7 +204,7 @@ import Foundation
     /// - returns: The `SentryId` of the event or `SentryId.empty` if the event is not sent.
     @objc(captureMessage:)
     @discardableResult public static func capture(message: String) -> SentryId {
-        return SentrySDKInternal.capture(message: message).sentryId
+        return SentrySDKInternal.capture(message: message)
     }
     
     /// Captures a message event and sends it to Sentry. Only the data in this scope object will be added
@@ -211,7 +214,7 @@ import Foundation
     /// - returns: The `SentryId` of the event or `SentryId.empty` if the event is not sent.
     @objc(captureMessage:withScope:)
     @discardableResult public static func capture(message: String, scope: Scope) -> SentryId {
-        return SentrySDKInternal.capture(message: message, scope: scope).sentryId
+        return SentrySDKInternal.capture(message: message, scope: scope)
     }
     
     /// Captures a message event and sends it to Sentry. Maintains the global scope but mutates scope
@@ -221,7 +224,7 @@ import Foundation
     /// - returns: The `SentryId` of the event or `SentryId.empty` if the event is not sent.
     @objc(captureMessage:withScopeBlock:)
     @discardableResult public static func capture(message: String, block: @escaping (Scope) -> Void) -> SentryId {
-        return SentrySDKInternal.capture(message: message, block: block).sentryId
+        return SentrySDKInternal.capture(message: message, block: block)
     }
     
     /// Captures user feedback that was manually gathered and sends it to Sentry.
@@ -388,12 +391,6 @@ import Foundation
         SentrySDKInternal.stopProfiler()
     }
     #endif
-}
-
-extension SentryIdWrapper {
-    var sentryId: SentryId {
-        SentryId(uuidString: sentryIdString)
-    }
 }
 
 // swiftlint:enable file_length