+++ Tests/SentryTests/SentryHubTests.swift
@@ -56,13 +56,13 @@ class SentryHubTests: XCTestCase {
             abnormalSession.environment = options.environment
         }
         
-        func getSut(withMaxBreadcrumbs maxBreadcrumbs: UInt = 100) -> SentryHub {
+        func getSut(withMaxBreadcrumbs maxBreadcrumbs: UInt = 100) -> SentryHubInternal {
             options.maxBreadcrumbs = maxBreadcrumbs
             return getSut(options)
         }
         
-        func getSut(_ options: Options, _ scope: Scope? = nil) -> SentryHub {
-            let hub = SentryHub(client: client, andScope: scope, andCrashWrapper: sentryCrashWrapper, andDispatchQueue: dispatchQueueWrapper)
+        func getSut(_ options: Options, _ scope: Scope? = nil) -> SentryHubInternal {
+            let hub = SentryHubInternal(client: client, andScope: scope, andCrashWrapper: sentryCrashWrapper, andDispatchQueue: dispatchQueueWrapper)
             hub.bindClient(client)
             return hub
         }
@@ -98,7 +98,7 @@ class SentryHubTests: XCTestCase {
         sentryOption.dsn = "https://6cc9bae94def43cab8444a99e0031c28@o447951.ingest.sentry.io/5428557"
                 
         let scope = Scope()
-        let sentryHub = SentryHub(client: SentryClient(options: sentryOption), andScope: scope)
+        let sentryHub = SentryHubInternal(client: SentryClientInternal(options: sentryOption), andScope: scope)
 
         let error = NSError(domain: "Test.CaptureErrorWithRealDSN", code: 12)
         sentryHub.capture(error: error)
@@ -186,7 +186,7 @@ class SentryHubTests: XCTestCase {
     }
     
     func testScopeEnriched_WithInitializer() {
-        let hub = SentryHub(client: nil, andScope: Scope())
+        let hub = SentryHubInternal(client: nil, andScope: Scope())
         XCTAssertFalse(hub.scope.contextDictionary.allValues.isEmpty)
         XCTAssertNotNil(hub.scope.contextDictionary["os"])
         XCTAssertNotNil(hub.scope.contextDictionary["device"])
@@ -201,7 +201,7 @@ class SentryHubTests: XCTestCase {
         let crashWrapper = SentryCrashWrapper(processInfoWrapper: processInfoWrapper)
         
         // Act
-        let hub = SentryHub(client: nil, andScope: Scope(), andCrashWrapper: crashWrapper, andDispatchQueue: TestSentryDispatchQueueWrapper())
+        let hub = SentryHubInternal(client: nil, andScope: Scope(), andCrashWrapper: crashWrapper, andDispatchQueue: TestSentryDispatchQueueWrapper())
 
         // Assert
         XCTAssertNil(hub.scope.contextDictionary["runtime"])
@@ -216,7 +216,7 @@ class SentryHubTests: XCTestCase {
         let crashWrapper = SentryCrashWrapper(processInfoWrapper: processInfoWrapper)
         
         // Act
-        let hub = SentryHub(client: nil, andScope: Scope(), andCrashWrapper: crashWrapper, andDispatchQueue: TestSentryDispatchQueueWrapper())
+        let hub = SentryHubInternal(client: nil, andScope: Scope(), andCrashWrapper: crashWrapper, andDispatchQueue: TestSentryDispatchQueueWrapper())
         
         // Assert
         let runtimeContext = try XCTUnwrap (hub.scope.contextDictionary["runtime"] as? [String: String])
@@ -234,7 +234,7 @@ class SentryHubTests: XCTestCase {
         let crashWrapper = SentryCrashWrapper(processInfoWrapper: processInfoWrapper)
         
         // Act
-        let hub = SentryHub(client: nil, andScope: Scope(), andCrashWrapper: crashWrapper, andDispatchQueue: TestSentryDispatchQueueWrapper())
+        let hub = SentryHubInternal(client: nil, andScope: Scope(), andCrashWrapper: crashWrapper, andDispatchQueue: TestSentryDispatchQueueWrapper())
 
         // Assert
         let runtimeContext = try XCTUnwrap (hub.scope.contextDictionary["runtime"] as? [String: String])
@@ -249,7 +249,7 @@ class SentryHubTests: XCTestCase {
     }
     
     func testScopeEnriched_WhenCreatingDefaultScope() {
-        let hub = SentryHub(client: nil, andScope: nil)
+        let hub = SentryHubInternal(client: nil, andScope: nil)
         
         let scope = hub.scope
         XCTAssertFalse(scope.contextDictionary.allValues.isEmpty)
@@ -276,11 +276,11 @@ class SentryHubTests: XCTestCase {
     }
     
     func testAddUserToTheScope() throws {
-        let client = SentryClient(
+        let client = SentryClientInternal(
             options: fixture.options,
             fileManager: fixture.fileManager
         )
-        let hub = SentryHub(client: client, andScope: Scope())
+        let hub = SentryHubInternal(client: client, andScope: Scope())
         
         let user = User()
         user.userId = "123"
@@ -574,7 +574,7 @@ class SentryHubTests: XCTestCase {
     
     func testSaveCrashTransaction_SavesTransaction() throws {
         let scope = fixture.scope
-        let sut = SentryHub(client: fixture.client, andScope: scope)
+        let sut = SentryHubInternal(client: fixture.client, andScope: scope)
         
         let transaction = sut.startTransaction(transactionContext: TransactionContext(name: fixture.transactionName, operation: fixture.transactionOperation, sampled: .yes, sampleRate: nil, sampleRand: nil))
 
@@ -590,7 +590,7 @@ class SentryHubTests: XCTestCase {
     func testSaveCrashTransaction_withSampleRateRand_SavesTransaction() throws {
         // Arrange
         let scope = fixture.scope
-        let sut = SentryHub(client: fixture.client, andScope: scope)
+        let sut = SentryHubInternal(client: fixture.client, andScope: scope)
         
         let transaction = sut.startTransaction(
             transactionContext: TransactionContext(
@@ -615,7 +615,7 @@ class SentryHubTests: XCTestCase {
     
     func testSaveCrashTransaction_NotSampled_DoesNotSaveTransaction() throws {
         let scope = fixture.scope
-        let sut = SentryHub(client: fixture.client, andScope: scope)
+        let sut = SentryHubInternal(client: fixture.client, andScope: scope)
         
         let transaction = sut.startTransaction(transactionContext: TransactionContext(name: fixture.transactionName, operation: fixture.transactionOperation, sampled: .no, sampleRate: nil, sampleRand: nil))
 
@@ -627,7 +627,7 @@ class SentryHubTests: XCTestCase {
 
     func testSaveCrashTransaction_NotSampledWithSampleRateRand_DoesNotSaveTransaction() throws {
         let scope = fixture.scope
-        let sut = SentryHub(client: fixture.client, andScope: scope)
+        let sut = SentryHubInternal(client: fixture.client, andScope: scope)
         
         let transaction = sut.startTransaction(
             transactionContext: TransactionContext(
@@ -1049,7 +1049,7 @@ class SentryHubTests: XCTestCase {
         sut = fixture.getSut()
         let options = fixture.options
         options.releaseName = nil
-        let client = SentryClient(options: options)
+        let client = SentryClientInternal(options: options)
         sut.bindClient(client)
         
         givenCrashedSession()
@@ -1144,7 +1144,7 @@ class SentryHubTests: XCTestCase {
         sut = fixture.getSut()
         let options = fixture.options
         options.releaseName = nil
-        let client = SentryClient(options: options)
+        let client = SentryClientInternal(options: options)
         sut.bindClient(client)
         
         // Act
@@ -1224,7 +1224,7 @@ class SentryHubTests: XCTestCase {
     }
     
     func testCaptureReplay() {
-        class SentryClientMockReplay: SentryClient {
+        class SentryClientMockReplay: SentryClientInternal {
             var replayEvent: SentryReplayEvent?
             var replayRecording: SentryReplayRecording?
             var videoUrl: URL?
@@ -1343,13 +1343,13 @@ class SentryHubTests: XCTestCase {
         }
     }
     
-    private func addBreadcrumbThroughConfigureScope(_ hub: SentryHub) {
+    private func addBreadcrumbThroughConfigureScope(_ hub: SentryHubInternal) {
         hub.configureScope({ scope in
             scope.addBreadcrumb(self.fixture.crumb)
         })
     }
     
-    private func captureConcurrentWithSession(count: Int, _ capture: @escaping (SentryHub) -> Void) {
+    private func captureConcurrentWithSession(count: Int, _ capture: @escaping (SentryHubInternal) -> Void) {
         let sut = fixture.getSut()
         sut.startSession()
         
@@ -1529,7 +1529,7 @@ class SentryHubTests: XCTestCase {
         fixture.currentDateProvider.setDate(date: fixture.currentDateProvider.date().addingTimeInterval(bySeconds))
     }
     
-    private func assert(withScopeBreadcrumbsCount count: Int, with hub: SentryHub) {
+    private func assert(withScopeBreadcrumbsCount count: Int, with hub: SentryHubInternal) {
         let scopeBreadcrumbs = hub.scope.serialize()["breadcrumbs"] as? [AnyHashable]
         XCTAssertNotNil(scopeBreadcrumbs)
         XCTAssertEqual(scopeBreadcrumbs?.count, count)