+++ Tests/SentryTests/SentryTests.m
@@ -38,7 +38,7 @@
             initWithDict:@{ @"dsn" : @"https://username:password@app.getsentry.com/12345" }
         didFailWithError:&error];
 
-    SentryClient *client = [[SentryClient alloc] initWithOptions:options];
+    SentryClientInternal *client = [[SentryClientInternal alloc] initWithOptions:options];
     XCTAssertNil(error);
     XCTAssertNil([SentrySDKInternal.currentHub getClient]);
     [SentrySDKInternal.currentHub bindClient:client];